Struct carboxyl_window::SourceWindow

pub struct SourceWindow<S> {
    // some fields omitted
}

A reactive window implementation generic over the event source.

Methods

impl<S> SourceWindow<S>

fn new(source: S) -> SourceWindow<S>

Create a new Glium loop.

Parameters

tick_length is the minimum duration of a tick in nanoseconds.

fn set_capture(&mut self, capture: Signal<bool>)

Mutably set cursor capturing signal

fn capture(self, capture: Signal<bool>) -> SourceWindow<S>

Provide cursor capturing signal